Split packer and unpacker
